+++ app/Haskell/DAP/GHCi/Utility.hs
@@ -4,13 +4,8 @@ module Haskell.DAP.GHCi.Utility where
 
 import qualified GHC
 import qualified GHCi.UI.Monad as G
-import HscTypes
 import Outputable
-import PprTyThing
-import Debugger
 import Exception
-import DynFlags
-import RtClosureInspect
 
 import qualified Data.Char as CH
 import qualified Data.ByteString as BS
@@ -18,12 +13,14 @@ import qualified Data.Text.Encoding as T
 import qualified Data.Text as T
 import qualified Data.List as L
 import qualified Text.Read as R
+import Data.Maybe
 import Data.Word
 
 import Control.Monad.IO.Class
+import Control.Concurrent
 
 import Haskell.DAP.GHCi.Constant
-import qualified GHCi.DAP.IFData as D
+import Haskell.DAP.GHCi.Type
 
 -- |
 --
@@ -123,65 +120,22 @@ printDAP dat = do
 
 -- |
 --
-showStoppedEventBody :: Maybe GHC.ExecResult -> G.GHCi ()
-showStoppedEventBody Nothing = return ()
-showStoppedEventBody (Just (GHC.ExecComplete { GHC.execResult = Right _ })) = do
-  let body = Right D.defaultStoppedEventBody {
-             D.reasonStoppedEventBody = "complete"
-           } :: Either String D.StoppedEventBody
-      outStr = _DAP_HEADER ++ showDAP body
-  liftIO $ putStrLn "[DAP][INFO] debug complete"
-  liftIO $ putStrLn outStr
-  
-showStoppedEventBody (Just (GHC.ExecComplete { GHC.execResult = Left (SomeException e)})) = do
-  let body = Right D.defaultStoppedEventBody {
-             D.reasonStoppedEventBody = "complete"
-           , D.descriptionStoppedEventBody = show e
-           , D.textStoppedEventBody = show e
-           } :: Either String D.StoppedEventBody
-      outStr = _DAP_HEADER ++ showDAP body
-  liftIO $ putStrLn $ "[DAP][INFO] debug complete with exception. " ++ show e
-  liftIO $ putStrLn outStr
+clearTmpDAPContext :: G.GHCi ()
+clearTmpDAPContext = do
+  mvarCtx <- G.dapContextGHCiState  <$> G.getGHCiState 
 
-showStoppedEventBody (Just GHC.ExecBreak{}) = getExceptionResume >>= \case
-  Nothing  -> stoppedByBreak
-  Just _  -> stoopedWithException
+  ctx <- liftIO $ takeMVar mvarCtx
+  liftIO $ putMVar mvarCtx ctx{
+      traceCmdExecResultDAPContext   = []
+    , doContinueExecResultDAPContext = []
+    , runStmtDeclExceptionDAPContext = []
+    }
 
-  where
-    stoopedWithException = do
-      evalBody <- getEvalBody "_exception" True
-      let body = Right D.defaultStoppedEventBody {
-                 D.reasonStoppedEventBody = "exception"
-               , D.descriptionStoppedEventBody = D.resultEvaluateBody evalBody
-               , D.textStoppedEventBody = D.resultEvaluateBody evalBody
-               } :: Either String D.StoppedEventBody
-          outStr = _DAP_HEADER ++ showDAP body
-      liftIO $ putStrLn "[DAP][INFO] break with exception."
-      liftIO $ putStrLn outStr
-
-    stoppedByBreak = do
-      let body = Right D.defaultStoppedEventBody {
-                 D.reasonStoppedEventBody = "step"
-               } :: Either String D.StoppedEventBody
-          outStr = _DAP_HEADER ++ showDAP body
-
-      liftIO $ putStrLn "[DAP][INFO] breaks."
-      liftIO $ putStrLn outStr
-
-
--- |
---
-getExceptionResume :: G.GHCi (Maybe GHC.Resume)
-getExceptionResume = GHC.getResumeContext >>= \case
-    [] -> return Nothing
-    (x:_) -> if isExceptionResume x then return (Just x)
-               else return Nothing
 
 -- |
 --
 isExceptionResume :: GHC.Resume -> Bool
-isExceptionResume r = L.isInfixOf "exception" (GHC.resumeDecl r)
-
+isExceptionResume (GHC.Resume{GHC.resumeBreakInfo = a}) = isNothing a
 
 
 -- |
@@ -195,67 +149,6 @@ parseNameErrorHandler e = liftIO $ print e >> return []
 showTermErrorHandler :: SomeException -> G.GHCi SDoc
 showTermErrorHandler e = return $ text $ show e
 
-
--- |
---
---
-getEvalBody :: String -> Bool -> G.GHCi D.EvaluateBody
-getEvalBody nameStr isForce =
-  gcatch (GHC.parseName nameStr) parseNameErrorHandler >>= withNames
-
-  where
-    withNames [] = return D.defaultEvaluateBody {
-                            D.resultEvaluateBody = "Not in scope: " ++ nameStr
-                          , D.typeEvaluateBody   = "force error."
-                          , D.variablesReferenceEvaluateBody = 0
-                          }
-    withNames (n:[]) = GHC.lookupName n >>= \case
-      Just ty -> withTyThing ty
-      Nothing -> return D.defaultEvaluateBody {
-                          D.resultEvaluateBody = "variable not found. " ++ nameStr
-                        , D.typeEvaluateBody   = "force error."
-                        , D.variablesReferenceEvaluateBody = 0
-                        }
-    withNames _ = return D.defaultEvaluateBody {
-                           D.resultEvaluateBody = "ambiguous name" ++ nameStr
-                         , D.typeEvaluateBody   = "force error."
-                         , D.variablesReferenceEvaluateBody = 0
-                         }
-
-    withTyThing (AnId i) = GHC.obtainTermFromId maxBound isForce i >>= withTerm i
-
-    withTyThing x = do
-      dflags <- getDynFlags
-      return D.defaultEvaluateBody {
-               D.resultEvaluateBody = "unsupported tything. " ++ showSDoc dflags (ppr x)
-             , D.typeEvaluateBody   = "force error."
-             , D.variablesReferenceEvaluateBody = 0
-             }
-
-    -- |
-    --  Term https://hackage.haskell.org/package/ghc-8.2.1/docs/RtClosureInspect.html
-    --
-    withTerm :: GHC.Id -> Term -> G.GHCi D.EvaluateBody
-    withTerm _ t@(Term ty _ _ _) = do
-      dflags <- getDynFlags
-      termSDoc <- gcatch (showTerm t) showTermErrorHandler
-      let typeStr = showSDoc dflags (pprTypeForUser ty)
-          valStr  = showSDoc dflags termSDoc
-
-      return D.defaultEvaluateBody {
-               D.resultEvaluateBody = valStr
-             , D.typeEvaluateBody   = typeStr
-             }
-
-    withTerm i _ = do
-      dflags <- getDynFlags
-      idSDoc <- pprTypeAndContents i
-      let (_, typeStr, valStr) = getNameTypeValue (showSDoc dflags idSDoc)
-      return D.defaultEvaluateBody {
-               D.resultEvaluateBody = valStr
-             , D.typeEvaluateBody  = typeStr
-             }
-
 -- |
 --
 getNameTypeValue :: String -> (String, String, String)
@@ -267,3 +160,17 @@ getNameTypeValue str = (strip nameStr, strip typeStr, strip valueStr)
     valueStr  = if elem "->" (words typeStr) then "function :: " ++ typeStr
                   else valueStr_
 
+
+-- |
+--
+getRunStmtSourceError :: G.GHCi String
+getRunStmtSourceError = do
+  mvarCtx <- G.dapContextGHCiState <$> G.getGHCiState 
+
+  ctx <- liftIO $ readMVar mvarCtx
+  let errs = runStmtDeclExceptionDAPContext ctx
+      msgs = "[DAP][ERROR] error occurred while runStmt." 
+            : map show errs
+
+  return $ L.intercalate "\n" msgs
+